added security authorised acces only

+++ code/app/Http/Controllers/HomeController.php
@@ -33,42 +33,80 @@ class HomeController extends Controller
     }
 
     public function addUser(Request $request){
-        $username = array($request->input('firstname'), $request->input('lastname'));
-        $email = array($request->input('firstname'),'.',$request->input('lastname'),'@email.com');
 
-        $query = DB::table('users')->insert([
-            'firstname'=>$request->input('firstname'),
-            'lastname'=>$request->input('lastname'),
-            'username'=>join("", $username),
-            'email'=>join("", $email),
-            'password'=> Hash::make('1234'),
-            'roleid'=>$request->input('role')
-        ]);
+        if(Auth::id()){
+            $roleID=Auth()->user()->roleid;
+            if($roleID==4){
+                $username = array($request->input('firstname'), $request->input('lastname'));
+                $email = array($request->input('firstname'),'.',$request->input('lastname'),'@email.com');
 
-        return back();
+                $query = DB::table('users')->insert([
+                    'firstname'=>$request->input('firstname'),
+                    'lastname'=>$request->input('lastname'),
+                    'username'=>join("", $username),
+                    'email'=>join("", $email),
+                    'password'=> Hash::make('1234'),
+                    'roleid'=>$request->input('role')
+                ]);
+
+                return back();
+            }
+            else{
+                abort(401);
+            }
+
+        }
+
+        
         
     }
 
     public function deleteUser($id){
-        $user = User::find($id);
-        $user->delete();
-        return back();
+
+        if(Auth::id()){
+            $roleID=Auth()->user()->roleid;
+            if($roleID==4){
+                $user = User::find($id);
+                $user->delete();
+                return back();            
+            }
+            else{
+                abort(401);
+            }
+        }
+
     }
 
     public function editUser($id){
-        $user = User::find($id);
+        if(Auth::id()){
+            $roleID=Auth()->user()->roleid;
+            if($roleID==4){
+                $user = User::find($id);
 
-        return view('editUser', compact('user'));
+                return view('editUser', compact('user'));
+            }
+            else{
+                abort(401);
+            }
+        }
     }
 
     public function updateUser(Request $request, $id){
-        $user = User::find($id);
-        $user->firstname = $request->input('firstname');
-        $user->lastname = $request->input('lastname');
-        $user->username = $request->input('username');
-        $user->roleid = $request->input('role');
-        $user->update();
-        return redirect('/account');
+        if(Auth::id()){
+            $roleID=Auth()->user()->roleid;
+            if($roleID==4){
+                $user = User::find($id);
+                $user->firstname = $request->input('firstname');
+                $user->lastname = $request->input('lastname');
+                $user->username = $request->input('username');
+                $user->roleid = $request->input('role');
+                $user->update();
+                return redirect('/account');
+            }
+            else{
+                abort(401);
+            }
+        }
     }
 
     public function index(){
